Move FFI definitions into the correct file.
These should match the underlying header location.
