+++ security-framework-sys/src/keychain.rs
@@ -1,7 +1,7 @@
-use core_foundation_sys::base::{Boolean, OSStatus, CFTypeID};
+use core_foundation_sys::base::{Boolean, OSStatus, CFTypeID, CFTypeRef};
 use libc::{c_char, c_void};
 
-use base::{SecAccessRef, SecKeychainRef};
+use base::{SecAccessRef, SecKeychainRef, SecKeychainItemRef};
 
 extern "C" {
     pub fn SecKeychainGetTypeID() -> CFTypeID;
@@ -23,4 +23,29 @@ extern "C" {
                              password: *const c_void,
                              usePassword: Boolean)
                              -> OSStatus;
+    #[cfg(target_os = "macos")]
+    pub fn SecKeychainFindGenericPassword(keychainOrArray: CFTypeRef,
+                                          serviceNameLength: u32,
+                                          serviceName: *const c_char,
+                                          accountNameLength: u32,
+                                          accountName: *const c_char,
+                                          passwordLength: *mut u32,
+                                          passwordData: *mut *mut u8,
+                                          itemRef: *mut SecKeychainItemRef)
+                                          -> OSStatus;
+    #[cfg(target_os = "macos")]
+    pub fn SecKeychainAddGenericPassword(keychain: SecKeychainRef,
+                                         serviceNameLength: u32,
+                                         serviceName: *const c_char,
+                                         accountNameLength: u32,
+                                         accountName: *const c_char,
+                                         passwordLength: u32,
+                                         passwordData: *const u8,
+                                         itemRef: *mut SecKeychainItemRef)
+                                         -> OSStatus;
+    #[cfg(target_os = "macos")]
+    // XXX First arg should be a SecKeychainAttributeList.
+    pub fn SecKeychainItemFreeContent(attrList: *const c_void,
+                                      data: *const c_void)
+                                      -> OSStatus;
 }
